requirements.txt
httpx==0.24.1
numpy==1.24.3
pandas==2.0.1
pyotp==2.9.0
PyQt5==5.15.9
PyYAML==6.0
Requests==2.31.0
typing_extensions==4.7.1
websocket_client==1.6.1
